+++ firmware/controllers/engine_cycle/spark_logic.h
@@ -7,11 +7,11 @@
 
 #pragma once
 
-void onTriggerEventSparkLogic(int rpm, efitick_t edgeTimestamp, float currentPhase, float nextPhase);
+void onTriggerEventSparkLogic(float rpm, efitick_t edgeTimestamp, float currentPhase, float nextPhase);
 void turnSparkPinHighStartCharging(IgnitionEvent *event);
 void fireSparkAndPrepareNextSchedule(IgnitionEvent *event);
 int getNumberOfSparks(ignition_mode_e mode);
 // fact: getInjectorDutyCycle is used by limpManager as cut reason but coil duty cycle is only logged not considered for control strategy
 // see also maxAllowedDwellAngle which only produces a warning without cutting spark
-percent_t getCoilDutyCycle(int rpm);
+percent_t getCoilDutyCycle(float rpm);
 void initializeIgnitionActions();
